More refactoring, broke up the rectange detction phase into two seperate stages, one to gather groups of approximately parallel lines and the other to search for pairs of groups approximately normal to each other and then generate rectangles from the lines contained in each group.

+++ src/algorithms/best_rectangle.cc
@@ -15,7 +15,7 @@ void BestRectangleSelector::apply(PipelineJob& job) {
         float thisScore = rectangles[index].getScore();
         if (thisScore > bestScore || index == 0) {
             bestScore = thisScore;
-            bestOrdinal = index;
+            bestOrdinal = static_cast<int>(index);
         }
     }
